size_t to int

+++ src/main.cpp
@@ -117,7 +117,7 @@ int main(int argc, char **argv)
     int i;
     int window_interval = dhcpmon_default_health_check_window;
     int max_unhealthy_count = dhcpmon_default_unhealthy_max_count;
-    size_t snaplen = dhcpmon_default_snaplen;
+    int snaplen = dhcpmon_default_snaplen;
     int make_daemon = 0;
     bool debug_mode = false;
     errno = 0;
